fixed StampDiscrete using non-zero pixel values for certain classes
Some classes don't initialize their pixels with the default constructor.
This caused StampDiscrete to sometimes compute absurd pixel values when
out of range (0), or between pixels (with bilinear interpolation).

The fix is a hack: instead of using the default constructor, an uninitialized
pixel stored in the static segment of the program is copied.

A better fix would be to have every Pixel be initialized at 0.