renewed the Stamper classes.
The Stamper classes now accept any type of images (although they may or
may not be appropriate for certain types of processing, for instance,
ImageRGBu8 is hardly appropriate for Gaussian noise).
Image types are given through a single template parameter.

Most Stamper parameters are now set after the construction via setters.
They now take a pointer to a Sampler and a const pointer to a Stamp.
Their Stamp shall not be modified, however the Sampler can and will be,
as they used PRNG generators.

StamperTexton now utilizes the Stamp interpolation rule instead of its own.
A user can thus modify it to change the behavior of the Texton Stamp.