Added StampDiscreteWithMask (in need to be tested).
This is a StampDiscrete holding a boolean mask of the size
of the underlying image. Accessing a pixel in pixel() will first
check if the same pixel(s) (with bilinear interpolation) is
set to true in the mask. The mask can be set with setMask()
and modified with a reference to it (function mask()).
It is allocated to the size of the image and set to true by default.

Additionnaly, the boundary checks are now all performed with
virtual protected function is_in_range(x, y).

Finally, StampDiscrete now holds a const reference to the image
instead of a copy of it. This means we are allowed to modify the
content of the stamp outside of it, and that holding a large
image is now much cheaper. This change was made to be able to
hold several stamps with the same image, but different bitmasks,
without taking up too much memory.