Added several new ways to mipmap a texture.
MipmapBooleanImage can be used to filter a boolean image,
with the merging rule for A and B being (A || B).

MipmapBitmask can be used to filter a bitmask image of any type,
with the merging rule for A and B being (A | B).

MipmapCEContent can be used to construct mipmaps out of
a Mipmap class of the content (whithin a fully sized image)
and a Mipmap class of the alpha of the patch (fully sized image).
The result is a Mipmap class of the content, taking the space
of the patch (it has the same content, but whithin a small space).
All of these additions were made to filter the content exchange algorithm.

Additionnaly, the function mipmap() now takes two unsigned ints.