Made Mipmap classes more versatile.
Mipmap classes can now be initialized without a first image and
don't produce a mipmap until asked to with the generate() function.
The class is supposed to detect when functions are called prior to
setting a texture or generating when either of these functions are
supposed to have been called.