Added ContentExchange::Patch class.
A patch is composed of an origin, giving the top left of the region,
and an alpha map, giving either 0 or 1 where the non-squared patch
is within a rectangular region.
The class was built with a mipmapping feature which provides the
class with tools to filter every patches. These patches are
filtered within their original alpha map (taking the entire input
texture size), and an origin as well as the alpha map for each
level of mipmapping can be obtained.