Added ContentExchange::Content for storing content.
A content is a texture data compressed within a rectangle.
This rectangle is supposed to be combined with its (patch) parent's
alpha map to give out one of the possible contents of this patch.

Just like its patch counterpart, the Content class is provided with
a mipmap support: contents are filtered from an image of the input
texture's size and content to be copied where the patch is supposed
to be.

MipmapCEContent, which is the class providing the mipmapping,
has also been moved in content.h and its behavior has been changed
to use the result of MipmapCEPatch.