Added ContentExchange::PatchProcessor, to articulate patches and inputs.
PatchProcessor uses input textures to compute a content exchange
process result. It is meant to compute patches, appropriate
alternative content for each, and be able to blend them together to
give the output.

So far, an user is only able to get patches by providing a colormap
which shows the patches. This colormap can be obtained with the older
patch exchange algorithm.
As for the contents, only the default content is given.
Finally, there is no generate function to get an output, or
probability function to yield an output.