Added cleaning step before generating contents.
We found out contents could not be filtered using an alpha mipmap
and an entire rectangular area, as it mixes two (or more)
different contents that may have nothing to do with the output.
The area is thus cleaned, as in every pixel which is not in the
given content is erased.
There is a bug where the "erasing" would not set the pixel to 0
in case of ImageRGBs, so we use static memory to fix it.