Modified Mipmap default filtering to work on Imageu8 classes.
Previously this was impossible, because of the entire range
of the structure being used, as well as the fact that divisions
could break on it.

Note that there could still be small mistakes being
carried (and increase) over several levels of mipmaps,
because of the rounding from float to char.