Added several tests (list bellow).
Test RPN is for Random Phase Noise by Example.
Test Gradient is to compute the gradient of a scalar map.
Test Height to texture creates a texture depending on a height map.
Test multi-texture Content Exchange is exactly what it sounds like.
Test terrain amplification amplifies a terrain with a sparse dictionary.