Added RegionFinder classes, TexturePool, and a matching extension for ContentExchange.
RegionFinder is a namespace used to find regions in a texture
that match a given regions and a metric.
It allows for any kind of region searching and is extremly versatile.

TexturePool is a simple class mixing textures, linear transformations,
and specific texture-changing functions to yield a pool of transformed textures.

The matching extension for ContentExchange is used to amplify a guidance
using a pool of exemplars.
It has been adapted to gradient maps for terrain synthesis and amplification.